feat: move fileReader share logic to new AbstractImportCommand.php
Add new blank DatabaseImportCommand.php
Renamed FileReaderImportConfig.php -> ImportConfig
