+++ elasticms-cli/src/Command/Import/AbstractImportCommand.php
@@ -1,0 +1,203 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\CLI\Command\Import;
+
+use App\CLI\Client\File\ImportConfig;
+use EMS\CommonBundle\Common\Admin\AdminHelper;
+use EMS\CommonBundle\Common\Command\AbstractCommand;
+use EMS\CommonBundle\Contracts\CoreApi\Endpoint\Data\DataInterface;
+use EMS\CommonBundle\Search\Search;
+use EMS\CommonBundle\Storage\File\FileInterface;
+use EMS\CommonBundle\Storage\NotFoundException;
+use EMS\CommonBundle\Storage\StorageManager;
+use EMS\Helpers\Standard\Hash;
+use EMS\Helpers\Standard\Json;
+use EMS\Helpers\Standard\UuidGenerator;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
+
+abstract class AbstractImportCommand extends AbstractCommand
+{
+    private const string ARGUMENT_CONTENT_TYPE = 'content-type';
+    private const string OPTION_CONFIG = 'config';
+    private const string OPTION_DRY_RUN = 'dry-run';
+    private const string OPTION_LIMIT = 'limit';
+    private const string OPTION_FLUSH_SIZE = 'flush-size';
+    private const string OPTION_MERGE = 'merge';
+
+    protected string $contentType;
+    protected bool $dryRun;
+    protected bool $merge;
+    protected int $flushSize;
+    protected ?int $limit = null;
+    private ExpressionLanguage $expressionLanguage;
+
+    public function __construct(
+        private readonly AdminHelper $adminHelper,
+        private readonly StorageManager $storageManager,
+    ) {
+        parent::__construct();
+    }
+
+    #[\Override]
+    protected function configure(): void
+    {
+        $this
+            ->addArgument(self::ARGUMENT_CONTENT_TYPE, InputArgument::REQUIRED, 'Content type target')
+            ->addOption(self::OPTION_CONFIG, null, InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY, 'Config(s) json, file path or hash', [])
+            ->addOption(self::OPTION_DRY_RUN, null, InputOption::VALUE_NONE, 'Just do a dry run')
+            ->addOption(self::OPTION_MERGE, null, InputOption::VALUE_REQUIRED, 'Perform a merge or replace', true)
+            ->addOption(self::OPTION_FLUSH_SIZE, null, InputOption::VALUE_REQUIRED, 'Flush size for the queue', 100)
+            ->addOption(self::OPTION_LIMIT, null, InputOption::VALUE_REQUIRED, 'Limit the rows')
+        ;
+    }
+
+    #[\Override]
+    protected function initialize(InputInterface $input, OutputInterface $output): void
+    {
+        parent::initialize($input, $output);
+        $this->contentType = $this->getArgumentString(self::ARGUMENT_CONTENT_TYPE);
+        $this->dryRun = $this->getOptionBool(self::OPTION_DRY_RUN);
+        $this->merge = $this->getOptionBool(self::OPTION_MERGE);
+        $this->flushSize = $this->getOptionInt(self::OPTION_FLUSH_SIZE);
+        $this->limit = $this->getOptionIntNull(self::OPTION_LIMIT);
+
+        $this->expressionLanguage = new ExpressionLanguage();
+    }
+
+    public function import(ImportConfig $config, \Generator $records): void
+    {
+        $coreApi = $this->adminHelper->getCoreApi();
+        $contentTypeApi = $coreApi->data($this->contentType);
+        if (!$coreApi->isAuthenticated()) {
+            throw new \RuntimeException(\sprintf('Not authenticated for %s, run ems:admin:login', $this->adminHelper->getCoreApi()->getBaseUrl()));
+        }
+        $ouuids = $config->deleteMissingDocuments ? $this->searchExistingOuuids() : [];
+
+        $progressBar = $this->io->createProgressBar();
+        $count = 0;
+        $queue = $coreApi->queue($this->flushSize)->addFlushCallback(fn () => $progressBar->advance());
+
+        foreach ($records as $row) {
+            $ouuid = $this->createOuuid($config, $row);
+
+            $rawData = $config->defaultData;
+            $rawData['_sync_metadata'] = $row;
+
+            if (null !== $ouuidVersionExpression = $config->ouuidVersionExpression) {
+                $rawData['_version_uuid'] = UuidGenerator::fromValue(
+                    value: $this->expressionLanguage->evaluate($ouuidVersionExpression, ['row' => $row])
+                );
+            }
+
+            if ($ouuid) {
+                unset($ouuids[$ouuid]);
+            }
+
+            if (!$this->dryRun) {
+                $queue->add($contentTypeApi->indexAsync(ouuid: $ouuid, rawData: $rawData, merge: $this->merge));
+            }
+
+            ++$count;
+        }
+
+        $queue->flush();
+        $progressBar->finish();
+        $this->io->newLine();
+
+        $notReadable = \count($records->getReturn());
+        if ($notReadable > 0) {
+            $this->io->warning(\sprintf('Could not read %d records', $notReadable));
+        }
+
+        if (!$this->dryRun && $config->deleteMissingDocuments && \count($ouuids) > 0) {
+            $this->deleteMissingDocuments($contentTypeApi, ...\array_keys($ouuids));
+        }
+
+        $this->io->definitionList(
+            'Summary',
+            ['Index' => $count],
+            ['Delete' => \count($ouuids)]
+        );
+    }
+
+    protected function getImportConfig(): ImportConfig
+    {
+        $inputs = $this->getOptionStringArray(self::OPTION_CONFIG, false);
+
+        $configs = \array_map(fn (string $input) => match (true) {
+            Json::isJson($input) => Json::decode($input),
+            default => Json::decode($this->getFile($input)->getContent()),
+        }, $inputs);
+
+        return ImportConfig::createFromArray(
+            config: \array_merge(...$configs)
+        );
+    }
+
+    protected function getFile(string $fileIdentifier): FileInterface
+    {
+        try {
+            return $this->storageManager->getFile($fileIdentifier);
+        } catch (NotFoundException) {
+            return $this->adminHelper->getCoreApi()->file()->getFile($fileIdentifier);
+        }
+    }
+
+    /**
+     * @param array<int, array<mixed>> $row
+     */
+    private function createOuuid(ImportConfig $config, array $row): ?string
+    {
+        if (null === $config->ouuidExpression) {
+            return null;
+        }
+
+        $ouuid = $this->expressionLanguage->evaluate($config->ouuidExpression, ['row' => $row]);
+        $prefix = $config->ouuidPrefix;
+
+        return match (true) {
+            $config->generateOuuid => (string) UuidGenerator::fromValue(($prefix ?? '').$ouuid),
+            null !== $prefix => Hash::string($prefix.$ouuid),
+            $config->generateHash => Hash::string(\sprintf('FileReaderImport:%s:%s', $this->contentType, $ouuid)),
+            default => $ouuid,
+        };
+    }
+
+    private function deleteMissingDocuments(DataInterface $api, string ...$ouuids): void
+    {
+        $this->io->newLine(2);
+        $this->io->section(\sprintf('%d documents have not been updated and will be deleted', \count($ouuids)));
+        $progressBar = $this->io->createProgressBar(\count($ouuids));
+        foreach ($ouuids as $ouuid) {
+            $api->delete($ouuid);
+            $progressBar->advance();
+        }
+        $progressBar->finish();
+        $this->io->newLine();
+    }
+
+    /**
+     * @return array<string, bool>
+     */
+    private function searchExistingOuuids(): array
+    {
+        $ouuids = [];
+        $search = new Search([
+            $this->adminHelper->getCoreApi()->meta()->getDefaultContentTypeEnvironmentAlias($this->contentType),
+        ]);
+        $search->setSources(['_id']);
+        $search->setContentTypes([$this->contentType]);
+
+        foreach ($this->adminHelper->getCoreApi()->search()->scroll($search) as $hit) {
+            $ouuids[$hit->getOuuid()] = true;
+        }
+
+        return $ouuids;
+    }
+}
